requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-openai
openai
faiss-cpu
psycopg2-binary
redis